pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codebase-markdown-visualizer"
version = "0.1.0"
description = "Transform a Python codebase into interlinked markdown files for Obsidian."
authors = [
  { name="Your Name", email="you@example.com" }
]
dependencies = [
  "PyYAML",
  # Add additional dependencies if needed
]
